pyproject.toml
[project]
name = "ust-agent-hub-mvp"
version = "0.1.0"
description = ""
authors = [
    {name = "287344_ust",email = "Djamel.Touati@ust.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "jsonschema (>=4.25.1,<5.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "requests (>=2.32.5,<3.0.0)"
]
